+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rusty_meta_threads"
-version = "0.3.0"
+version = "0.4.0"
 edition = "2021"
 description = "Community Rust SDK for integrating with Meta Threads API"
 repository = "https://github.com/Thesephi/rusty_meta_threads.git"
