+++ src/mod.rs
@@ -2,5 +2,6 @@ pub mod auth;
 pub mod profiles;
 pub mod reply_management;
 pub mod retrieve_media;
+pub use crate::shared::{MetaMediaResponse, Paging, ThreadsApiRespErrorPayload};
 mod shared;
 mod utils;
