+++ src/profiles.rs
@@ -1,5 +1,16 @@
-use crate::shared::ThreadsUserProfile;
+use crate::shared::ThreadsApiRespErrorPayload;
 use log::debug;
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug)]
+pub struct ThreadsUserProfile {
+    pub id: String,
+    pub username: Option<String>,
+    pub name: Option<String>,
+    pub threads_profile_picture_url: Option<String>,
+    pub threads_biography: Option<String>,
+    pub error: Option<ThreadsApiRespErrorPayload>,
+}
 
 // @TODO have fields as fn arguments instead of hardcoding
 pub async fn get_profile_info(bearer_token: &str) -> Result<ThreadsUserProfile, reqwest::Error> {
@@ -20,7 +31,7 @@ pub async fn get_profile_info(bearer_token: &str) -> Result<ThreadsUserProfile, 
             debug!("failed to retrieve Threads user profile: {:#?}", error);
             // @TODO consider using Err instead of Ok
             Ok(ThreadsUserProfile {
-                id: None,
+                id: String::from(""),
                 username: None,
                 name: None,
                 threads_biography: None,
