+++ src/shared.rs
@@ -1,45 +1,16 @@
 use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
-pub struct ThreadsUserProfile {
-    pub id: Option<String>,
-    pub username: Option<String>,
-    pub name: Option<String>,
-    pub threads_profile_picture_url: Option<String>,
-    pub threads_biography: Option<String>,
-    pub error: Option<ThreadsApiRespErrorPayload>,
-}
-
-// https://developers.facebook.com/docs/threads/reply-management#a-thread-s-conversations
-#[derive(Deserialize, Debug)]
-pub struct MetaMedia {
-    pub id: String,
-    pub is_reply_owned_by_me: Option<bool>,
-    pub username: Option<String>,
-    pub text: Option<String>,
-    pub timestamp: Option<String>,
-    pub media_product_type: Option<String>, // THREADS | ...
-    pub media_type: Option<String>,         // TEXT_POST | ...
-    pub media_url: Option<String>,
-    pub permalink: Option<String>,
-    pub shortcode: Option<String>,
-    pub has_replies: Option<bool>,
-    // pub root_post: { id: "1234567890" },
-    // pub replied_to: { id: "1234567890" },
-    // pub is_reply: bool,
-    // pub hide_status: String, // NOT_HUSHED | ...
-}
-
-#[derive(Deserialize, Debug)]
-pub struct MetaMediaResponse {
-    pub data: Option<Vec<MetaMedia>>,
+pub struct MetaMediaResponse<T> {
+    pub data: Option<Vec<T>>,
     pub paging: Option<Paging>,
-    error: Option<ThreadsApiRespErrorPayload>,
+    pub error: Option<ThreadsApiRespErrorPayload>,
 }
 
 #[derive(Deserialize, Debug)]
 pub struct Paging {
     cursors: Cursors,
+    pub next: String,
 }
 
 #[derive(Deserialize, Debug)]
